requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
